requirements.txt
gradio  # HF Spaces pins this to 6.7.0; do not pin here
python-dotenv
openai
chromadb
requests
trafilatura
pdfplumber
huggingface_hub
litellm
lxml_html_clean
neo4j>=5.14.0
anthropic>=0.21.0

# dev / testing
pytest

pandas
numpy
streamlit
